+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partitioner/KeyGroupStreamPartitioner.java
@@ -15,13 +15,12 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.runtime.partitioner;
+package org.apache.flink.runtime.io.network.partitioner;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.api.java.functions.KeySelector;
 import org.apache.flink.runtime.plugable.SerializationDelegate;
 import org.apache.flink.runtime.state.KeyGroupRangeAssignment;
-import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 import org.apache.flink.util.Preconditions;
 
 /**
@@ -51,14 +50,14 @@ public class KeyGroupStreamPartitioner<T, K> extends StreamPartitioner<T> implem
 
 	@Override
 	public int[] selectChannels(
-		SerializationDelegate<StreamRecord<T>> record,
+		SerializationDelegate<T> record,
 		int numberOfOutputChannels) {
 
 		K key;
 		try {
-			key = keySelector.getKey(record.getInstance().getValue());
+			key = keySelector.getKey(record.getInstance());
 		} catch (Exception e) {
-			throw new RuntimeException("Could not extract key from " + record.getInstance().getValue(), e);
+			throw new RuntimeException("Could not extract key from " + record.getInstance(), e);
 		}
 		returnArray[0] = KeyGroupRangeAssignment.assignKeyToParallelOperator(key, maxParallelism, numberOfOutputChannels);
 		return returnArray;
