+++ flink-streaming-java/src/test/java/org/apache/flink/runtime/io/network/partitioner/ForwardPartitionerTest.java
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.flink.runtime.io.network.partitioner;
+
+import org.apache.flink.api.java.tuple.Tuple;
+import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
+
+import org.junit.Test;
+
+/**
+ * Tests for {@link ForwardPartitioner}.
+ */
+public class ForwardPartitionerTest extends StreamPartitionerTest {
+
+	@Override
+	public StreamPartitioner<StreamRecord<Tuple>> createPartitioner() {
+		return new ForwardPartitioner<>();
+	}
+
+	@Test
+	public void testSelectChannelsInterval() {
+		assertSelectedChannel(0, 1);
+		assertSelectedChannel(0, 2);
+		assertSelectedChannel(0, 1024);
+	}
+}
