+++ flink-streaming-java/src/test/java/org/apache/flink/runtime/io/network/partitioner/GlobalPartitionerTest.java
@@ -15,32 +15,27 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.runtime.partitioner;
+package org.apache.flink.runtime.io.network.partitioner;
 
 import org.apache.flink.api.java.tuple.Tuple;
+import org.apache.flink.streaming.runtime.streamrecord.StreamRecord;
 
 import org.junit.Test;
 
-import static org.junit.Assert.assertTrue;
-
 /**
- * Tests for {@link RebalancePartitioner}.
+ * Tests for {@link GlobalPartitioner}.
  */
-public class RebalancePartitionerTest extends StreamPartitionerTest {
+public class GlobalPartitionerTest extends StreamPartitionerTest {
 
 	@Override
-	public StreamPartitioner<Tuple> createPartitioner() {
-		return new RebalancePartitioner<>();
+	public StreamPartitioner<StreamRecord<Tuple>> createPartitioner() {
+		return new GlobalPartitioner<>();
 	}
 
 	@Test
-	public void testSelectChannelsInterval() {
-		int initialChannel = streamPartitioner.selectChannels(serializationDelegate, 3)[0];
-		assertTrue(0 <= initialChannel);
-		assertTrue(3 > initialChannel);
-
-		assertSelectedChannel((initialChannel + 1) % 3, 3);
-		assertSelectedChannel((initialChannel + 2) % 3, 3);
-		assertSelectedChannel((initialChannel + 3) % 3, 3);
+	public void testSelectChannels() {
+		assertSelectedChannel(0, 1);
+		assertSelectedChannel(0, 2);
+		assertSelectedChannel(0, 1024);
 	}
 }
